# pump_failure_service/openapi.yaml
openapi: 3.0.3
info:
  title: Pump Failure Prediction API
  version: 1.0.0
  description: |
    Servei Python per predicció de fallades de bombes d'aigua basat en dades de sensors.
    Inclou endpoints per predicció en temps real, reentrenament del model i consulta d'estat.

servers:
  - url: http://localhost:8000
    description: Local server

paths:
  /predict:
    post:
      summary: Predir probabilitat de fallada
      description: Rep dades en temps real i retorna la probabilitat i el diagnòstic de fallada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vibration
                - power
                - temperature
                - rpm
                - pressure
                - flow
                - uptime
              properties:
                vibration:
                  type: number
                  format: float
                  example: 0.12
                power:
                  type: number
                  format: float
                  example: 350.5
                temperature:
                  type: number
                  format: float
                  example: 55.3
                rpm:
                  type: integer
                  example: 1450
                pressure:
                  type: number
                  format: float
                  example: 2.3
                flow:
                  type: number
                  format: float
                  example: 12.5
                uptime:
                  type: number
                  format: float
                  example: 3600
      responses:
        '200':
          description: Resultat de la predicció
          content:
            application/json:
              schema:
                type: object
                properties:
                  failure_probability:
                    type: number
                    format: float
                    example: 0.85
                  prediction_label:
                    type: string
                    example: "FAILURE"
                  model_version:
                    type: string
                    example: "v1.3.6"
        '400':
          description: Dades d'entrada invàlides

  /train:
    post:
      summary: Reentrenar el model
      description: Entrena el model amb noves dades i guarda el fitxer versionat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                training_data:
                  type: array
                  items:
                    type: object
                    properties:
                      vibration: { type: number, format: float }
                      power: { type: number, format: float }
                      temperature: { type: number, format: float }
                      rpm: { type: integer }
                      pressure: { type: number, format: float }
                      flow: { type: number, format: float }
                      uptime: { type: number, format: float }
                      failure: { type: boolean }
              example:
                training_data:
                  - vibration: 0.15
                    power: 360
                    temperature: 54
                    rpm: 1450
                    pressure: 2.5
                    flow: 12.1
                    uptime: 3700
                    failure: false
      responses:
        '200':
          description: Resultat de l'entrenament
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "success" }
                  model_version: { type: string, example: "v1.3.6" }
                  model_file: { type: string, example: "/models/pump_failure_model_v1.3.6.pkl" }
                  trained_on_records: { type: integer, example: 10523 }
                  training_duration_seconds: { type: number, format: float, example: 45.7 }
                  message: { type: string, example: "Model retrained successfully" }

  /status:
    get:
      summary: Consultar estat del model
      description: Retorna informació del model actiu, mètriques i fitxer associat.
      responses:
        '200':
          description: Estat del model
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_version: { type: string, example: "v1.3.6" }
                  model_file: { type: string, example: "/models/pump_failure_model_v1.3.6.pkl" }
                  last_training: { type: string, format: date-time, example: "2025-08-12T10:00:00Z" }
                  records_seen: { type: integer, example: 125000 }
                  accuracy: { type: number, format: float, example: 0.94 }
                  f1_score: { type: number, format: float, example: 0.92 }